Feature/data table (#12)
* Installed library associated with expandable panels

Includes code for the component to display this information (manually inserted).

* Cards can be rendered by extracting relevant info from array

Instead of having replicated code for each individual card, each card  renders information according to the current element of the array. So, a template was created for each card.

* Implemented Cards to be displayed in a drop down menu

Each individual node card will be displayed in the drop down menu. Once it is clicked the card will expand displaying new information. Also removed a now unnecessary file.

* Completed a Code Tidy Up

Renaming classes, variables and removing unnecessary code. Formatted the 'dummy' data to give a better idea of what a card will look like.

* Configured props for node information to move between components

Information (with manual input at this stage) can be transferred from the dashboard component to the drop down node list using props. The manual input needs to be updated to the nodes that the user selects.

* Fixed the code regarding a warning that appeared

There was a red warning that appeared with iterating through the props. Have fixed this error.

* Commented code and adjusted some prop configurations

Placed the drop down menu on the same dashboard component as the chart. Have also made some minor adjustments with respect to demonstrating an example of props and data movement between components.

* Made adjustments according to the feedback from my pull request

Used eslint to fix formatting issues and changed the directory structure for the components I was working on.

* Fixed some more eslint errors but still need fix the props error

* Fixed the prop types mapping eslint error

* Changed input values in the array to integers only

When we extract information from the API calls, chances are it will be only numerical values. So updated the formatting to accomodate this.