Extracting the data values processing & displaying totals (#37)
* Changed layout of the dashboard to look more presentable.

* Had to update the way the node panels were displayed

Due to the previous Node Data Table covering the graph, I have fixed it with a new component.

* Obtaining the Props information from DashBoard Panel for Node Display

Connected the checklistToggleMap information to the PanelSet Component to obtain the nodes the user selected.

* Formatted & Commented the Code.

* Made minor changes to suppress some console warnings

* Managed to extract node specific data according to the model

For this commit it is only displaying the sum of the value associated to a node, but can process them further (e.g. minimum, maximum). Need to implement  functionality to deal with null values.

* Tidied & Commented code.

Need to: Able to handle null values. determine what input values there are.

* Ability to handle null values & round values to 2dp

* Commits prior to merge - pull from develop

* Added extra calculations for the values (i.e. max & average)

* Placed the information of each node into the expandable card component

* Extracting Data Parsed into the Node Information Segment

* Fixed bug between label & data received

The labels and corresponding data values did not match elements, since the node in (labels[0]) would map to nodeTimeDataValues[0] which was just a string of the date-time. So this was incrementing the nodeTimeDataValues element by one.

* Added labels associated with the particular 'value'

* Added the scroll bar.

* Code clean up, including file removal and comments

* Updated node data font & imports according to PR feedback.